✨ Feat : 프록시와 연관관계
